Fix "tooltips shown in a notification alter its layout"
By default, "tooltip()" adds the tooltip element to the element in which
the function is called. If a "container" option is provided the tooltip
is added instead to the given element, although visually it will be
shown for the element in which the function is called.

The tooltips shown in notifications, when shown near the bottom, could
cause a scroll bar to appear in the notification wrapper, which in turn
could change the layout of the notification and cause a "jumpy" effect.
Now the tooltip is added to the notification wrapper, which prevents the
scroll bar from appearing when the tooltip is shown (even if the tooltip
overflows the notification wrapper).

This also fixes a wrong HTML markup, as the tooltip is shown with a
"div" element, but it was added to a "span" element; now it is added to
another "div" element.

Fixes #100

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>