Do not send push notifications when nothing was deleted
The iOS App always sent an API request when it deleted a notification.
By accident this was also done when it did the delete caused by a delete-push
Until now the notificiations app always sent a delete-push in case of an API
call. Meaning this circled all the time.
So now we only send a push if something was actually deleted.

Signed-off-by: Joas Schilling <coding@schilljs.com>