Fix push notifications for multibyte notifications
json_encode is messing with multibyte characters a lot,
replacing them with something along "\u1234" instead of the assumed 2-times overhead.
The data length in our encryption is a hard limit, but we don't want to
make non-utf8 subjects unnecessary short. So this function tries to cut
it off first.
If that is not enough it always cuts off 5 characters in multibyte-safe
fashion until the json_encode-d string is shorter then the given limit.

Signed-off-by: Joas Schilling <coding@schilljs.com>