Also check responses of 400 errors and so delete unknown devices
So before this PR, we tried to act on 400 errors already,
but I didn't know that they throw an exception on the `post()`,
so instead of performing on the response body,
we `continue;` after we logged an info :hushed:

This is of course not what was intended. Further more,
since the notifications are grouped and the proxy had an error
which would immediately bail out with a 400 when there was 1 broken notification,
instead of sending the rest, 1 broken notification was enough to not
send any notification anymore. And since we never deleted the broken
device the instance didn't "heal" itself, but always broke like this.

Signed-off-by: Joas Schilling <coding@schilljs.com>