Prepend * to document title when tab hidden and new notifications
When a new notification arrives, if the tab is hidden, and we didn't
already change the title, we prepend a * to the title to make pinned
tabs show a changed-highlight and unpinned tabs show a visible change.

When the tab becomes visible again we revert the change, unless the
title was already changed by Talk or any other app.

Ref: https://github.com/nextcloud/notifications/issues/892 #892