Remove the * we prepend to title if notifications count goes down.
We reset the title if the notifications are cleared in another tab/window.

This doesn't make the pinned-tab's highlight in firefox go away though,
because changing the title back is just more chaning-the-title.
If anything it would increase the amount of highting. Now we've
changed it twice.

Can't find a way to tell firefox to un-highlight the tab.

Chrome doesn't seem to do the highlighting thing anyway.

Ref https://github.com/nextcloud/notifications/issues/892 #892
https://github.com/nextcloud/notifications/pull/894