Fix (visually) removing first notification when executing an action
When an action is executed first its element is faded out and then the
notification data is removed, which removes the component as a whole.
However, the index of the notification was not provided, so the event
that handles the "remove" event always removed the data of the first
notification. Due to this when the action was triggered in another
notification the first notification was also (visually) removed (and
only until the notifications were fetched again from the server).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>